Allow the pressure expression to set the filename of the Hypre solver output files. Append those output filenames with the simulation time, rkstage, and patch ID. Also allow the pressure to be cleaved from its children now that we fixed the timestep in the SetCurrentTime expression.

+++ src/CCA/Components/Wasatch/Expressions/BasicExprBuilder.cc
@@ -1007,7 +1007,7 @@ namespace Wasatch{
       
       exprParams->get("Values", initialMoments, nEqs); // get only the first nEqs moments
       
-      assert(initialMoments.size() == nEqs);
+      assert( (int) initialMoments.size() == nEqs );
       
       const int nMoments = initialMoments.size();
       typedef Expr::ConstantExpr<SVolField>::Builder Builder;
